backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-dotenv==1.0.1
supabase==2.10.0
cryptography==44.0.0
pydantic==2.10.4
pydantic-settings==2.12.0
python-multipart==0.0.20
websockets==14.1
bcrypt==5.0.0
